+++ Shop/Areas/Administrator/Controllers/AspNetUsersController.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Net;
 using System.Web;
 using System.Web.Mvc;
+using Shop.Areas.Administrator.Data.message;
 using Shop.EF;
 
 namespace Shop.Areas.Administrator.Controllers
@@ -17,91 +18,144 @@ namespace Shop.Areas.Administrator.Controllers
         // GET: Administrator/AspNetUsers
         public ActionResult Index()
         {
-            return View(db.AspNetUsers.ToList());
+            if (Session["taikhoanadmin"] == null)
+            {
+                return RedirectToAction("Error401", "MainPage");
+            }
+            else
+            {
+                return View(db.AspNetUsers.ToList());
+            }
         }
 
         // GET: Administrator/AspNetUsers/Details/5
         public ActionResult Details(string id)
         {
-            if (id == null)
+            if (Session["taikhoanadmin"] == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Error401", "MainPage");
             }
-            AspNetUser aspNetUser = db.AspNetUsers.Find(id);
-            if (aspNetUser == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                AspNetUser aspNetUser = db.AspNetUsers.Find(id);
+                if (aspNetUser == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aspNetUser);
             }
-            return View(aspNetUser);
         }
 
         // GET: Administrator/AspNetUsers/Create
         public ActionResult Create()
         {
-            return View();
+            if (Session["taikhoanadmin"] == null)
+            {
+                return RedirectToAction("Error401", "MainPage");
+            }
+            else
+            {
+                return View();
+            }
         }
 
         // POST: Administrator/AspNetUsers/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
+        [ValidateInput(false)]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Id,Email,EmailConfirmed,PasswordHash,SecurityStamp,PhoneNumber,PhoneNumberConfirmed,TwoFactorEnabled,LockoutEndDateUtc,LockoutEnabled,AccessFailedCount,UserName,ngaysinh,profile,avatar,hoten,diachi")] AspNetUser aspNetUser)
         {
-            if (ModelState.IsValid)
+            if (Session["taikhoanadmin"] == null)
             {
-                db.AspNetUsers.Add(aspNetUser);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Error401", "MainPage");
             }
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    db.AspNetUsers.Add(aspNetUser);
+                    db.SaveChanges();
+                    Notification.set_flash("Thêm mới tài khoản đăng nhập thành công !", "success");
+                    return RedirectToAction("Index");
+                }
 
-            return View(aspNetUser);
+                return View(aspNetUser);
+            }
         }
 
         // GET: Administrator/AspNetUsers/Edit/5
         public ActionResult Edit(string id)
         {
-            if (id == null)
+            if (Session["taikhoanadmin"] == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Error401", "MainPage");
             }
-            AspNetUser aspNetUser = db.AspNetUsers.Find(id);
-            if (aspNetUser == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                AspNetUser aspNetUser = db.AspNetUsers.Find(id);
+                if (aspNetUser == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aspNetUser);
             }
-            return View(aspNetUser);
         }
 
         // POST: Administrator/AspNetUsers/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
+        [ValidateInput(false)]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Id,Email,EmailConfirmed,PasswordHash,SecurityStamp,PhoneNumber,PhoneNumberConfirmed,TwoFactorEnabled,LockoutEndDateUtc,LockoutEnabled,AccessFailedCount,UserName,ngaysinh,profile,avatar,hoten,diachi")] AspNetUser aspNetUser)
         {
-            if (ModelState.IsValid)
+            if (Session["taikhoanadmin"] == null)
             {
-                db.Entry(aspNetUser).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Error401", "MainPage");
             }
-            return View(aspNetUser);
+            else
+            {
+                if (ModelState.IsValid)
+                {
+                    db.Entry(aspNetUser).State = EntityState.Modified;
+                    db.SaveChanges();
+                    Notification.set_flash("Cập nhật tài khoản thành công !", "success");
+                    return RedirectToAction("Index");
+                }
+                return View(aspNetUser);
+            }
         }
 
         // GET: Administrator/AspNetUsers/Delete/5
         public ActionResult Delete(string id)
         {
-            if (id == null)
+            if (Session["taikhoanadmin"] == null)
             {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                return RedirectToAction("Error401", "MainPage");
             }
-            AspNetUser aspNetUser = db.AspNetUsers.Find(id);
-            if (aspNetUser == null)
+            else
             {
-                return HttpNotFound();
+                if (id == null)
+                {
+                    return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+                }
+                AspNetUser aspNetUser = db.AspNetUsers.Find(id);
+                if (aspNetUser == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(aspNetUser);
             }
-            return View(aspNetUser);
         }
 
         // POST: Administrator/AspNetUsers/Delete/5
@@ -109,10 +163,18 @@ namespace Shop.Areas.Administrator.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(string id)
         {
-            AspNetUser aspNetUser = db.AspNetUsers.Find(id);
-            db.AspNetUsers.Remove(aspNetUser);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["taikhoanadmin"] == null)
+            {
+                return RedirectToAction("Error401", "MainPage");
+            }
+            else
+            {
+                AspNetUser aspNetUser = db.AspNetUsers.Find(id);
+                db.AspNetUsers.Remove(aspNetUser);
+                db.SaveChanges();
+                Notification.set_flash("Xóa tài khoản Account thành công !", "success");
+                return RedirectToAction("Index");
+            }
         }
 
         protected override void Dispose(bool disposing)
